FilterDialog works ok now. Unfilter by selecting a greater weight value. The code is quite ugly though...
git-svn-id: https://svn.code.sf.net/p/socnetv/code@317 fc2d5e9b-b3ce-4657-a66d-b27cdf764933

+++ trunk/src/filteredgesdialog.h
@@ -41,7 +41,7 @@ public:
 public slots:
 	void gatherData ();
 signals:
-	void userChoices( double, bool);	
+	void userChoices( float, bool);	
 private:
 	Ui::FilterEdgesDialog ui;
